// one-cms/src/api/openapi/models/SpaceGroup.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Server API - Login
 * The Restful APIs of Login.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { SpaceGroupSpacesInner } from './SpaceGroupSpacesInner';
import {
  SpaceGroupSpacesInnerFromJSON,
  SpaceGroupSpacesInnerFromJSONTyped,
  SpaceGroupSpacesInnerToJSON,
} from './SpaceGroupSpacesInner';

/**
 *
 * @export
 * @interface SpaceGroup
 */
export interface SpaceGroup {
  /**
   *
   * @type {string}
   * @memberof SpaceGroup
   */
  spaceGroupId?: string;
  /**
   *
   * @type {string}
   * @memberof SpaceGroup
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof SpaceGroup
   */
  description?: string;
  /**
   *
   * @type {string}
   * @memberof SpaceGroup
   */
  thumbnail?: string;
  /**
   *
   * @type {string}
   * @memberof SpaceGroup
   */
  status?: SpaceGroupStatusEnum;
  /**
   *
   * @type {Date}
   * @memberof SpaceGroup
   */
  startAt?: Date;
  /**
   *
   * @type {Date}
   * @memberof SpaceGroup
   */
  endAt?: Date;
  /**
   *
   * @type {Array<SpaceGroupSpacesInner>}
   * @memberof SpaceGroup
   */
  spaces?: Array<SpaceGroupSpacesInner>;
}

/**
 * @export
 */
export const SpaceGroupStatusEnum = {
  Enabled: 'enabled',
  Disabled: 'disabled',
} as const;
export type SpaceGroupStatusEnum = (typeof SpaceGroupStatusEnum)[keyof typeof SpaceGroupStatusEnum];

/**
 * Check if a given object implements the SpaceGroup interface.
 */
export function instanceOfSpaceGroup(value: object): boolean {
  let isInstance = true;

  return isInstance;
}

export function SpaceGroupFromJSON(json: any): SpaceGroup {
  return SpaceGroupFromJSONTyped(json, false);
}

export function SpaceGroupFromJSONTyped(json: any, ignoreDiscriminator: boolean): SpaceGroup {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    spaceGroupId: !exists(json, 'space_group_id') ? undefined : json['space_group_id'],
    name: !exists(json, 'name') ? undefined : json['name'],
    description: !exists(json, 'description') ? undefined : json['description'],
    thumbnail: !exists(json, 'thumbnail') ? undefined : json['thumbnail'],
    status: !exists(json, 'status') ? undefined : json['status'],
    startAt: !exists(json, 'start_at') ? undefined : new Date(json['start_at']),
    endAt: !exists(json, 'end_at') ? undefined : new Date(json['end_at']),
    spaces: !exists(json, 'spaces')
      ? undefined
      : (json['spaces'] as Array<any>).map(SpaceGroupSpacesInnerFromJSON),
  };
}

export function SpaceGroupToJSON(value?: SpaceGroup | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    space_group_id: value.spaceGroupId,
    name: value.name,
    description: value.description,
    thumbnail: value.thumbnail,
    status: value.status,
    start_at: value.startAt === undefined ? undefined : value.startAt.toISOString(),
    end_at: value.endAt === undefined ? undefined : value.endAt.toISOString(),
    spaces:
      value.spaces === undefined
        ? undefined
        : (value.spaces as Array<any>).map(SpaceGroupSpacesInnerToJSON),
  };
}